requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
google-genai>=1.0.0
python-dotenv==1.0.0
opencv-python-headless>=4.8.0
python-multipart>=0.0.9
numpy>=1.24.0
